Move metrics package to org.springframework.core
See gh-24878
