+++ spring-context/src/main/java/org/springframework/context/ConfigurableApplicationContext.java
@@ -21,10 +21,10 @@ import java.io.Closeable;
 import org.springframework.beans.BeansException;
 import org.springframework.beans.factory.config.BeanFactoryPostProcessor;
 import org.springframework.beans.factory.config.ConfigurableListableBeanFactory;
-import org.springframework.beans.metrics.ApplicationStartup;
 import org.springframework.core.env.ConfigurableEnvironment;
 import org.springframework.core.env.Environment;
 import org.springframework.core.io.ProtocolResolver;
+import org.springframework.core.metrics.ApplicationStartup;
 import org.springframework.lang.Nullable;
 
 /**
@@ -90,7 +90,7 @@ public interface ConfigurableApplicationContext extends ApplicationContext, Life
 
 	/**
 	 * Name of the {@link ApplicationStartup} bean in the factory.
-	 * @since 5.3.0
+	 * @since 5.3
 	 */
 	String APPLICATION_STARTUP_BEAN_NAME = "applicationStartup";
 
@@ -139,13 +139,13 @@ public interface ConfigurableApplicationContext extends ApplicationContext, Life
 	 * <p>This allows the application context to record metrics
 	 * during startup.
 	 * @param applicationStartup the new context event factory
-	 * @since 5.3.0
+	 * @since 5.3
 	 */
 	void setApplicationStartup(ApplicationStartup applicationStartup);
 
 	/**
 	 * Return the {@link ApplicationStartup} for this application context.
-	 * @since 5.3.0
+	 * @since 5.3
 	 */
 	ApplicationStartup getApplicationStartup();
