+++ spring-core/src/main/java/org/springframework/core/metrics/jfr/package-info.java
@@ -1,9 +1,9 @@
 /**
- * Support package for recording metrics during application startup.
+ * Support package for recording startup metrics using Java Flight Recorder.
  */
 @NonNullApi
 @NonNullFields
-package org.springframework.beans.metrics;
+package org.springframework.core.metrics.jfr;
 
 import org.springframework.lang.NonNullApi;
 import org.springframework.lang.NonNullFields;
